cleanup webpacks
