+++ scripts/start.js
@@ -4,7 +4,7 @@ const createWebpackServer = require('webpack-httpolyglot-server');
 
 const argv = require('minimist')(process.argv.slice(2));
 
-const config = require(`../webpack/${argv.env}.config`);
+const config = require(`../webpack/devConfig`);
 
 tasks.replaceWebpack();
 console.log('[Copy assets]');
@@ -16,7 +16,7 @@ console.log('-'.repeat(80));
 console.log('If you\'re developing Inject page,');
 console.log('please allow `https://localhost:3000` connections in Google Chrome,');
 console.log('and load unpacked extensions with `./dev` folder. (see https://developer.chrome.com/extensions/getstarted#unpacked)\n');
-createWebpackServer(config, {
+createWebpackServer(config.baseDevConfig(argv.env), {
   host: 'localhost',
   port: 3000
 });
